+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - improve display of computed multipage images
 - batch mode output is lazy
 - ++, join_sep, difference are lazy
+- removed trace system
 
 ## 9.0.13 2025/09/20